+++ meta_package_manager/bar_plugin/meta_package_manager.7h.py
@@ -9,7 +9,9 @@
 # <xbar.abouturl>https://github.com/kdeldycke/meta-package-manager</xbar.abouturl>
 # <xbar.var>boolean(VAR_SUBMENU_LAYOUT=false): Group packages into a sub-menu for each manager.</xbar.var>
 # <xbar.var>boolean(VAR_TABLE_RENDERING=false): Aligns package names and versions in a table for easier visual parsing.</xbar.var>
-# <swiftbar.environment>['VAR_SUBMENU_LAYOUT': false, 'VAR_TABLE_RENDERING': false]</swiftbar.environment>
+# <xbar.var>string(VAR_DEFAULT_FONT=""): Default font to use for non-monospaced.</xbar.var>
+# <xbar.var>string(VAR_MONOSPACE_FONT="font=Menlo size=12"): Default font for monospace fonts, uncluding errors.</xbar.var>
+# <swiftbar.environment>['VAR_SUBMENU_LAYOUT': false, 'VAR_TABLE_RENDERING': false, 'VAR_DEFAULT_FONT': '', 'VAR_MONOSPACE_FONT': 'font=Menlo size=12']</swiftbar.environment>
 
 """Xbar and SwiftBar plugin for Meta Package Manager (i.e. the :command:`mpm` CLI).
 
@@ -31,83 +33,16 @@ https://github.com/swiftbar/SwiftBar/issues/160
 
 import os
 import re
-import subprocess
 import sys
 from configparser import RawConfigParser
 from shutil import which
+from subprocess import run
 from unittest.mock import patch
 
 if sys.version_info >= (3, 8):
     from functools import cached_property
 else:
     cached_property = property
-
-
-def getenv_str(var, default=None):
-    """Utility to get environment variables.
-
-    Note that all environment variables are strings. Always returns a lowered-case
-    string.
-    """
-    value = os.environ.get(var, None)
-    if value is None:
-        return default
-    return str(value).lower()
-
-
-def getenv_bool(var, default=False):
-    """Utility to normalize boolean environment variables.
-
-    Relies on ``configparser.RawConfigParser.BOOLEAN_STATES`` to translate strings into boolean. See:
-    https://github.com/python/cpython/blob/89192c46da7b984811ff3bd648f8e827e4ef053c/Lib/configparser.py#L597-L599
-    """
-    value = getenv_str(var)
-    if value is None:
-        return default
-    return RawConfigParser.BOOLEAN_STATES[value]
-
-
-SUBMENU_LAYOUT = getenv_bool("VAR_SUBMENU_LAYOUT", False)
-""" Group packages into manager sub-menus.
-
-If ``True``, will replace the default flat layout with an alternative structure
-where actions are grouped into submenus, one for each manager.
-"""
-
-
-TABLE_RENDERING = getenv_bool("VAR_TABLE_RENDERING", True)
-""" Aligns package names and versions, like a table, for easier visual parsing.
-
-If ``True``, will aligns all items using a fixed-width font.
-"""
-
-# Make it easier to change font, sizes and colors of the output
-MONOSPACE = "font=Menlo size=12"
-FONTS = {
-    "normal": "",  # Use default system font
-    "summary": "",  # Package summary
-    "package": "",  # Indiviual packages
-    "error": f"{MONOSPACE} color=red",  # Errors
-}
-# Use a monospaced font on table-like rendering.
-if TABLE_RENDERING:
-    FONTS.update(dict.fromkeys(["summary", "package"], MONOSPACE))
-
-
-MPM_MIN_VERSION = (5, 0, 0)
-"""Mpm v5.0.0 was the first version taking care of the layout rendering."""
-
-
-IS_SWIFTBAR = getenv_bool("SWIFTBAR")
-"""SwiftBar is kind enough to tell us about its presence."""
-
-
-DARK_MODE = (
-    getenv_str("OS_APPEARANCE", "light") == "dark"
-    if IS_SWIFTBAR
-    else getenv_bool("XBARDarkMode")
-)
-"""Detect dark mode."""
 
 
 class MPMPlugin:
@@ -119,6 +54,9 @@ class MPMPlugin:
         - https://github.com/matryer/xbar-plugins/blob/main/CONTRIBUTING.md#plugin-api
         - https://github.com/swiftbar/SwiftBar#plugin-api
     """
+
+    mpm_min_version = (5, 0, 0)
+    """Mpm v5.0.0 was the first version taking care of the complete layout rendering."""
 
     @staticmethod
     def extended_environment():
@@ -149,6 +87,67 @@ class MPMPlugin:
             )
         )
         return env_copy
+
+    @staticmethod
+    def getenv_str(var, default=None):
+        """Utility to get environment variables.
+
+        Note that all environment variables are strings. Always returns a lowered-case
+        string.
+        """
+        value = os.environ.get(var, None)
+        if value is None:
+            return default
+        return str(value).lower()
+
+    @staticmethod
+    def getenv_bool(var, default=False):
+        """Utility to normalize boolean environment variables.
+
+        Relies on ``configparser.RawConfigParser.BOOLEAN_STATES`` to translate strings into boolean. See:
+        https://github.com/python/cpython/blob/89192c46da7b984811ff3bd648f8e827e4ef053c/Lib/configparser.py#L597-L599
+        """
+        value = MPMPlugin.getenv_str(var)
+        if value is None:
+            return default
+        return RawConfigParser.BOOLEAN_STATES[value]
+
+    @staticmethod
+    def normalize_params(font_string, valid_ids={"color", "font", "size"}):
+        valid_params = {}
+        for param in font_string.split():
+            param_id, param_value = param.split("=", 1)
+            if param_id in valid_ids:
+                valid_params[param_id] = param_value
+        return " ".join(f"{k}={v}" for k, v in valid_params.items())
+
+    @cached_property
+    def table_rendering(self):
+        """ Aligns package names and versions, like a table, for easier visual parsing.
+
+        If ``True``, will aligns all items using a fixed-width font.
+        """
+        return self.getenv_bool("VAR_TABLE_RENDERING", True)
+
+    @cached_property
+    def default_font(self):
+        """Make it easier to change font, sizes and colors of the output."""
+        return self.normalize_params(self.getenv_bool("VAR_DEFAULT_FONT", ""))
+
+    @cached_property
+    def monospace_font(self):
+        """Make it easier to change font, sizes and colors of the output."""
+        return self.normalize_params(self.getenv_bool("VAR_MONOSPACE_FONT", "font=Menlo size=12"))
+
+    @cached_property
+    def error_font(self):
+        """Error font is based on monospace font"""
+        return self.normalize_params(f"{self.monospace_font} color=red")
+
+    @cached_property
+    def is_swiftbar(self):
+        """SwiftBar is kind enough to tell us about its presence."""
+        return self.getenv_bool("SWIFTBAR")
 
     @staticmethod
     def locate_bin(*bin_names):
@@ -200,21 +199,20 @@ class MPMPlugin:
         MPMPlugin.pp("❗️", "dropdown=false")
         print("---")
 
-    @staticmethod
-    def print_error(message, submenu=""):
+    def print_error(self, message, submenu=""):
         """Print a formatted error line by line.
 
         A red, fixed-width font is used to preserve traceback and exception layout.
         """
         # Cast to string as we might directly pass exceptions for rendering.
         for line in str(message).strip().splitlines():
-            MPMPlugin.pp(
+            self.pp(
                 f"{submenu}{line}",
-                FONTS["error"],
+                self.error_font,
                 "trim=false",
                 "ansi=false",
                 "emojize=false",
-                "symbolize=false" if IS_SWIFTBAR else "",
+                "symbolize=false" if self.is_swiftbar else "",
             )
 
     def print_menu(self):
@@ -223,9 +221,7 @@ class MPMPlugin:
         code = None
         error = None
         try:
-            process = subprocess.run(
-                (*self.mpm_exec, "--version"), capture_output=True, encoding="utf-8"
-            )
+            process = run((*self.mpm_exec, "--version"), capture_output=True, encoding="utf-8")
             code = process.returncode
             error = process.stderr
         except FileNotFoundError as ex:
@@ -243,7 +239,7 @@ class MPMPlugin:
                 .groupdict()["version"]
             )
             mpm_version = tuple(map(int, version_string.split(".")))
-            if mpm_version >= MPM_MIN_VERSION:
+            if mpm_version >= self.mpm_min_version:
                 mpm_up_to_date = True
 
         if not mpm_installed or not mpm_up_to_date:
@@ -251,7 +247,7 @@ class MPMPlugin:
             self.print_error(error)
             print("---")
             action_msg = "Install" if not mpm_installed else "Upgrade"
-            min_version_str = ".".join(map(str, MPM_MIN_VERSION))
+            min_version_str = ".".join(map(str, self.mpm_min_version))
             self.pp(
                 f"{action_msg} mpm >= v{min_version_str}",
                 f"shell={self.python_path}",
@@ -264,29 +260,19 @@ class MPMPlugin:
                 # https://github.com/swiftbar/SwiftBar/issues/308
                 # Fallback to the only version that is working on SwiftBar.
                 f'param5=\\"meta-package-manager>={min_version_str}\\"',
-                FONTS["error"],
+                self.error_font,
                 "refresh=true",
                 "terminal=true",
             )
             return
 
         # Force a sync of all local package databases.
-        subprocess.run((*self.mpm_exec, "--verbosity", "ERROR", "sync"))
+        run((*self.mpm_exec, "--verbosity", "ERROR", "sync"))
 
         # Fetch outdated packages from all package managers available on the system.
         # We defer all rendering to mpm itself so it can compute more intricate layouts.
-        process = subprocess.run(
-            (
-                *self.mpm_exec,
-                "--verbosity",
-                "ERROR",
-                "outdated",
-                "--plugin",
-                "swiftbar" if IS_SWIFTBAR else "xbar",
-                "--plugin-submenu-layout" if SUBMENU_LAYOUT else "--plugin-flat-layout",
-                "--plugin-aligned-columns" if TABLE_RENDERING else "--plugin-no-alignment",
-                "--plugin-dark-mode" if DARK_MODE else "--plugin-light-mode",
-            ),
+        process = run(
+            (*self.mpm_exec, "--verbosity", "ERROR", "outdated", "--plugin-output"),
             capture_output=True,
             encoding="utf-8",
         )